Create the initial fbuild.bff to compile our lib's Objects
There are three elements to this fbuild.bff.
The first is the Compiler('gcc'),
telling FASTBuild where to find our compiler and its family

The next is ObjectList('Fib-Obj'),
which defines the source files, the options,
and the output path for object files.

The final element is Alias('all') which sets the default targets for FASTBuild
if none are specified

Each of these "functions" create nodes in FASTBuild's dependency graph.
A node's deps can be queried by running `fbuild -showdeps <NODENAME>`.
For most deps, a large number of the elements are dynamic.
Try running a build then run `fbuild -showdeps <NODENAME>`

TODO: make a shared/dynamic lib