Create ObjectList for main's objects
The ObjectList('Main-Obj') is almost identical to ObjectList('Fib-Obj')
Notice that we were able to move the Compiler-related parameters to a shared spot
for all the ObjectList nodes.
The nodes will implicitly pull this information in if they are in scope.

TODO: Link the DLL and main object into an executable