Use -O1 for Debug builds of FibIter
Originally in the toolchains, the .Config_$OS$ set .CompilerOption.
This was updated to .CompilerOption_OS
which is used to create .CompilerOption
when merging OS and flavour configs.

This pattern allows consumers of a .BuildConfig to keep track of
where different config came from.
This allows them to override choice parts of the config.
In this case, FibIter overrode the flavour config for Debug